work in progress, created abstract class and changes async
